requirements.txt
aiohttp==3.7.4.post0
alabaster==0.7.12
async-timeout==3.0.1
attrs==21.4.0
Babel==2.9.1
beautifulsoup4==4.10.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
colorama==0.4.4
discord-components==2.1.2
discord.py==1.7.3
docutils==0.17.1
furo==2021.11.23
idna==3.3
imagesize==1.3.0
Jinja2==3.0.3
livereload==2.6.3
MarkupSafe==2.0.1
multidict==5.2.0
packaging==21.3
Pillow==9.0.0
Pygments==2.11.1
pyparsing==3.0.6
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.1
Sphinx==4.3.2
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tornado==6.1
typing-extensions==4.0.1
urllib3==1.26.7
yarl==1.7.2
